refactor: share Task wire contracts
