+++ packages/db/src/board-contract.test.ts
@@ -18,7 +18,7 @@ test("the browser-safe board contract imports only types", () => {
   const typeSpecifiers = [
     ...source.matchAll(/(?:^|\n)\s*import\s+type\b[^;]*?from\s+"([^"]+)"/gu),
   ].map((match) => match[1]);
-  assert.deepEqual(typeSpecifiers, ["@prisma/client"]);
+  assert.deepEqual(typeSpecifiers, ["@prisma/client", "./wire-contract.js"]);
 });
 
 test("the package publishes the board contract as an isolated subpath", () => {
